Add support to print result of GET commands on wpa_cli console
Add support to print result of all driver "GET" commands on
wpa_cli console.

wpa_supplicant's callback API for the DRIVER command over
ctrl_iface (wpa_supplicant_driver_cmd) overwrites the
response_buffer if the return value is 0. This will not allow
the response buffer data to be logged on to wpa_cli console.

Hence return the strlen of response_buffer for the successful
ioctl calls of "GET" commands.

CRs-Fixed: 2763719
Change-Id: I5be2b8b4c37d56637948aa783cddd803dcd3b05d